Move repository creation to a separate package
