+++ server/application/src/main/kotlin/io/vitalir/kotlinhub/server/app/feature/repository/routes/RepositoryRoutes.kt
@@ -1,21 +1,12 @@
 package io.vitalir.kotlinhub.server.app.feature.repository.routes
 
-import arrow.core.Either
-import io.ktor.http.*
-import io.ktor.server.application.*
-import io.ktor.server.request.*
 import io.ktor.server.routing.*
-import io.vitalir.kotlinhub.server.app.common.routes.ResponseData
-import io.vitalir.kotlinhub.server.app.common.routes.extensions.respondWith
 import io.vitalir.kotlinhub.server.app.common.routes.jwtAuth
-import io.vitalir.kotlinhub.server.app.feature.repository.domain.model.CreateRepositoryData
-import io.vitalir.kotlinhub.server.app.feature.repository.domain.usecase.CreateRepositoryUseCase
+import io.vitalir.kotlinhub.server.app.feature.repository.routes.create.createRepositoryRoute
 import io.vitalir.kotlinhub.server.app.feature.repository.routes.getrepository.getRepositoryRoute
 import io.vitalir.kotlinhub.server.app.feature.repository.routes.remove.removeRepositoryForCurrentUserRoute
 import io.vitalir.kotlinhub.server.app.feature.repository.routes.update.updateRepositoryForCurrentUserRoute
-import io.vitalir.kotlinhub.server.app.infrastructure.auth.userId
 import io.vitalir.kotlinhub.server.app.infrastructure.di.AppGraph
-import io.vitalir.kotlinhub.shared.common.network.Url
 
 internal fun Routing.repositoryRoutes(
     repositoryGraph: AppGraph.RepositoryGraph,
@@ -26,57 +17,6 @@ internal fun Routing.repositoryRoutes(
         jwtAuth {
             removeRepositoryForCurrentUserRoute(repositoryGraph.removeRepositoryUseCase)
             updateRepositoryForCurrentUserRoute(repositoryGraph.updateRepositoryUseCase)
-        }
-    }
-}
-
-private fun Route.createRepositoryRoute(
-    createRepositoryUseCase: CreateRepositoryUseCase,
-) {
-    jwtAuth {
-        post {
-            val userId = call.userId
-            val request = call.receive<CreateRepositoryRequest>()
-            val createRepositoryData = CreateRepositoryData(
-                ownerId = userId,
-                name = request.name,
-                accessMode = request.accessMode,
-                description = request.description,
-            )
-
-            val result = createRepositoryUseCase(createRepositoryData)
-            val responseData = result.toCreateRepositoryResponseData()
-            call.respondWith(responseData)
-        }
-    }
-}
-
-private fun Either<CreateRepositoryUseCase.Error, Url>.toCreateRepositoryResponseData(): ResponseData {
-    return when (this) {
-        is Either.Left -> value.toResponseData()
-        is Either.Right -> ResponseData(
-            code = HttpStatusCode.Created,
-            body = CreateRepositoryResponse(repositoryUrl = value.toString()),
-        )
-    }
-}
-
-private fun CreateRepositoryUseCase.Error.toResponseData(): ResponseData {
-    return when (this) {
-        is CreateRepositoryUseCase.Error.UserDoesNotExist -> {
-            ResponseData.fromErrorData(
-                code = HttpStatusCode.Unauthorized,
-                errorMessage = "unauthorized",
-            )
-        }
-        is CreateRepositoryUseCase.Error.RepositoryAlreadyExists -> {
-            ResponseData.fromErrorData(
-                code = HttpStatusCode.BadRequest,
-                errorMessage = "repository $repositoryName already exists",
-            )
-        }
-        is CreateRepositoryUseCase.Error.Unknown -> {
-            ResponseData.serverError()
         }
     }
 }
