+++ server/application/src/main/kotlin/io/vitalir/kotlinhub/server/app/feature/repository/routes/getrepository/GetRepositoryResponse.kt
@@ -4,6 +4,6 @@ import io.vitalir.kotlinhub.server.app.feature.repository.routes.ApiRepository
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class GetRepositoryResponse(
+internal data class GetRepositoryResponse(
     val repository: ApiRepository,
 )
